Dfine one route for project steps
